+++ ui/src/plugin/pages/FailedImagesPage.tsx
@@ -6,6 +6,6 @@ import { FailedImages } from '../../pages/MirrorTargets/FailedImages';
 setFetchImpl((url, init) => consoleFetch(url, init));
 setApiBaseUrl('/api/proxy/plugin/oc-mirror-operator/resourceapi');
 
-const FailedImagesPage: React.FC<any> = (props) => <FailedImages {...props} />;
+const FailedImagesPage: React.FC<Record<string, unknown>> = (props) => <FailedImages {...props} />;
 
 export default FailedImagesPage;