+++ ui/src/plugin/pages/ImageSetDetailPage.tsx
@@ -6,6 +6,6 @@ import { ImageSetDetail } from '../../pages/ImageSets/ImageSetDetail';
 setFetchImpl((url, init) => consoleFetch(url, init));
 setApiBaseUrl('/api/proxy/plugin/oc-mirror-operator/resourceapi');
 
-const ImageSetDetailPage: React.FC<any> = (props) => <ImageSetDetail {...props} />;
+const ImageSetDetailPage: React.FC<Record<string, unknown>> = (props) => <ImageSetDetail {...props} />;
 
 export default ImageSetDetailPage;